basic cross platform executor
